Fix usage of links and routing
