+++ src/views/CatalogueOperations/i18n.json
